[refactor] 리뷰 유저 멤버 디렉토링 moved
