+++ src/main/java/potatoes/server/chat/entity/Chat.java
@@ -15,7 +15,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import potatoes.server.entity.Travel;
-import potatoes.server.entity.User;
+import potatoes.server.user.entity.User;
 
 @Getter
 @NoArgsConstructor(access = PROTECTED)
