+++ src/main/java/potatoes/server/chat/entity/ChatMessage.java
@@ -18,7 +18,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import potatoes.server.config.BaseTimeEntity;
-import potatoes.server.entity.User;
+import potatoes.server.user.entity.User;
 
 @Getter
 @NoArgsConstructor(access = PROTECTED)
