+++ src/main/java/potatoes/server/chat/entity/ChatUser.java
@@ -12,7 +12,7 @@ import jakarta.persistence.ManyToOne;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import potatoes.server.entity.User;
+import potatoes.server.user.entity.User;
 
 @Getter
 @NoArgsConstructor(access = PROTECTED)
