+++ src/main/java/potatoes/server/dto/MessageSubscribe.java
@@ -3,7 +3,7 @@ package potatoes.server.dto;
 import java.util.List;
 
 import potatoes.server.chat.entity.ChatMessage;
-import potatoes.server.entity.User;
+import potatoes.server.user.entity.User;
 import potatoes.server.utils.time.DateTimeUtils;
 
 public record MessageSubscribe(
