+++ src/main/java/potatoes/server/review/controller/ReviewController.java
@@ -1,4 +1,4 @@
-package potatoes.server.controller;
+package potatoes.server.review.controller;
 
 import java.util.List;
 import java.util.Optional;
