+++ src/main/java/potatoes/server/review/entity/Review.java
@@ -1,4 +1,4 @@
-package potatoes.server.entity;
+package potatoes.server.review.entity;
 
 import static jakarta.persistence.FetchType.*;
 import static jakarta.persistence.GenerationType.*;
@@ -18,6 +18,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import potatoes.server.config.BaseTimeEntity;
+import potatoes.server.entity.Travel;
+import potatoes.server.user.entity.User;
 
 @Getter
 @NoArgsConstructor(access = PROTECTED)
