+++ src/main/java/potatoes/server/review/entity/ReviewImage.java
@@ -1,4 +1,4 @@
-package potatoes.server.entity;
+package potatoes.server.review.entity;
 
 import static jakarta.persistence.GenerationType.*;
 import static lombok.AccessLevel.*;
