+++ src/main/java/potatoes/server/user/repository/UserRepository.java
@@ -1,10 +1,10 @@
-package potatoes.server.repository;
+package potatoes.server.user.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import potatoes.server.entity.User;
+import potatoes.server.user.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findByEmail(String email);
